% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_agg_date_span.R
\name{aggregate_by_date_span}
\alias{aggregate_by_date_span}
\title{Aggregate Met Data from date Span}
\usage{
aggregate_by_date_span(df, df_met, progress = "none")
}
\arguments{
\item{df}{dataframe containing span range}

\item{df_met}{dataframe containg met data}

\item{progress}{supply "time" for progress bar}
}
\description{
Given averaged data with span ranges in "date_start" and "date_end "columns and met data with "date" column, 
average met data over span range and append to data frame
}
