% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_agg_date_span_wrk.R
\name{aggregate_by_date_span_worker}
\alias{aggregate_by_date_span_worker}
\title{Aggregate Met Data from date Span}
\usage{
aggregate_by_date_span_worker(df, df_met)
}
\arguments{
\item{df}{dataframe containing span range}

\item{df_met}{dataframe containg met data}
}
\description{
Given averaged data with span ranges in "date_start" and "date_end "columns and met data with "date" column, 
average met data over span range and append to data frame
passed from wrapper function line by line via adply()
}
