% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_aircraft_cal_flags.R
\name{aircraft_cal_flags}
\alias{aircraft_cal_flags}
\title{Aircraft Cal Flags}
\usage{
aircraft_cal_flags(d)
}
\arguments{
\item{d}{Raw AQD NOx output}
}
\description{
When given a dataframe of AQD NOx aircraft data (designed using MOYA Data) locates calibrations,
asks the user whether they are good or bad cals, then, using this information, recalculates the concentrations using a linear drift
}
