% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_aircraft_clean_negatvies.R
\name{aircraft_clean_negatives}
\alias{aircraft_clean_negatives}
\title{Aircraft Clean Negative}
\usage{
aircraft_clean_negatives(d)
}
\arguments{
\item{d}{Output of aircraft_cal_flags}
}
\description{
Removes Negatives from the adjusted concentraions from NOx (AQD) aircraft data
}
