% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_average_by_time.R
\name{average_by_time}
\alias{average_by_time}
\title{Average By Time}
\usage{
average_by_time(
  df,
  time_stamp = c("floor", "round", "ceiling")[1],
  res = "1 min",
  .f = mean,
  time_col = "date",
  ...
)
}
\arguments{
\item{df}{data.frame}

\item{time_stamp}{"floor" "round" or "ceiling". How is the time_stamp rounded to the resolution? 
conside the implication on the resulting timestamp. Floor is default, giving a timestamp that referes to the begining
of the averaging period}

\item{res}{time resolution to round to. Use lubridate "unit" format. Default "1 min"}

\item{.f}{function to use when using summarise_all. default mean}

\item{time_col}{what column contains the time information, default date}

\item{...}{extra arguments to pass to .f}
}
\description{
Function used to average data to new time resolution using the dplyr grouping. It is an attempt to be more efficent than
openair::timeAverage to be used when dealing with unusually large datasets. It is certainly nowhere near as robust. \cr
Use at your own risk
}
\author{
W. S. Drysdale
}
