% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_backtrajconfirm.R
\name{backTrajconfirm}
\alias{backTrajconfirm}
\title{Back Trajectory Confirm}
\usage{
backTrajconfirm(backTraj, pol, start, end, s)
}
\arguments{
\item{backTraj}{A backTrajectory file, pre merged with concentration data}

\item{pol}{pollutants within dataset}

\item{start}{string in format "dd/mm/yyyy" to be passed to selectBydate()}

\item{end}{string in format "dd/mm/yyyy" to be passed to selectBydate()}
}
\description{
Confirms with the user to continue plotting Back Trajectories, as these can take time
}
