% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_btparam.R
\name{bt_fastnox_param_monitor}
\alias{bt_fastnox_param_monitor}
\title{BT Tower Parameter Monitor}
\usage{
bt_fastnox_param_monitor(d)
}
\arguments{
\item{d}{Dataframe containing the raw NOx_crit output file - or a group of the bound together}
}
\description{
Creates a PDF of key paramaters on a time plot. Saves in current working directory
}
