% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btconfig.R
\docType{data}
\name{btfilepath}
\alias{btfilepath}
\title{BT Tower config}
\format{An object of class \code{character} of length 1.}
\usage{
btfilepath
}
\description{
Contains variables for the location of the BT tower time series
}
\keyword{datasets}
