% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_btadddata_fastnox.R
\name{btnox_adddata2}
\alias{btnox_adddata2}
\title{BT Tower Add Data 2}
\usage{
btnox_adddata2(main_series, o3_new_data, nox_new_data, version)
}
\arguments{
\item{main_series}{dataframe formatted as above}

\item{o3_new_data}{raw output from envirolgger when selecting o3 conc only}

\item{nox_new_data}{crit output file from AQD NOx instrument}

\item{version}{supply version number of data}
}
\value{
dataframe with new CSV reformatted and appended to the end, with QC flags applied
}
\description{
Adds data in the raw  output from WACLs Ozone box and Fast NOx at the BT Tower to main CSV files (main_series) in a specific format
columns for main_series: date: converted to POSIXct using ymd_hms()
                         no: NO / ppb
                         no2: NO2 / ppb
                         nox: Total NOx / ppb
                         o3: O3 / ppb
                         qc_flag_no: QC Flag 0-4
                         qc_flag_no2: QC Flag 0-4
                         qc_flag_nox: QC Flag 0-4
                         qc_flag_o3: QC Flag 0-4
Instrument name, location and version number are currently controlled by editing this function, where this infomation is passed to ncas_filename()
but could be adapted to have these as arguments in the main function.
}
