% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_btadddata_fastnox_met.R
\name{btnox_adddata2_met}
\alias{btnox_adddata2_met}
\title{BT Fast NOx add met data}
\usage{
btnox_adddata2_met(new_met, old_met, version)
}
\arguments{
\item{new_met}{raw dataframe from AQD NOx}

\item{old_met}{ongoing met file}
}
\value{
New Dataframe with new met appended at the end
}
\description{
Takes raw AQD NOx output and appends the met data to an ongoing file
}
