% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_BTT_cal_clean.R
\name{BTT_cal_clean}
\alias{BTT_cal_clean}
\title{BT Tower Calibration Clean}
\usage{
BTT_cal_clean(d, percent_drift)
}
\arguments{
\item{d}{dataframe of bound crit files}

\item{percent_drift}{numerical value of percentage drift allowed i.e 50 percent is provided as 50}
}
\description{
As the BTT calibrations performed automatically, and are less likley to be incorrect than an Aircraft Calibration, the user
is not asked whether they are good or bad on a per cal basis. This function provides some broad error catching for bad cals.
If the calibration parameters drift outside of a set percetange away from the average *nominal*, they are converted to NA
}
