% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_BTT_calibration_profile.R
\name{BTT_calibration_profile}
\alias{BTT_calibration_profile}
\title{BTT Calibraion Profile}
\usage{
BTT_calibration_profile(
 
    output_file_path = "C:\\\\Users\\\\Will\\\\Google Drive\\\\PhD\\\\site_BT Tower\\\\AQD_Calibration_Profile",
  crit_file_folder = "1_min_crit_files",
  percentage_drift = 50
)
}
\arguments{
\item{output_file_path}{Location containing the folder where all of the crit files are stored}

\item{crit_file_folder}{name of folder located within output_file_path contining crit files}

\item{percentage_drift}{to be supplied to BTT_cal_clean to limit drift of bad cals}
}
\description{
From the 1 min AQD crit file, a profile of the corrected calibration paramaters is produced
}
\author{
Will S. Drysdale
}
