% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_BTT_CEH_data_re_preprocess.R
\name{BTT_CEH_data_reprocess}
\alias{BTT_CEH_data_reprocess}
\title{CEH_data_reprocess}
\usage{
BTT_CEH_data_reprocess(d, header)
}
\arguments{
\item{d}{dataframe of raw data}

\item{header}{dataframe containing the headers for d as its own column headers}
}
\description{
Alternative preprocess script to supply the 2012-13 data to new flux code.
Built ontop of the orginal script by Adam R. Vaughan for BTT(2012) and the Beijing Site(2017)
}
\author{
Will S. Drysdale
}
