% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_BTT_clean_high.R
\name{BTT_clean_high}
\alias{BTT_clean_high}
\title{BTT Clean High}
\usage{
BTT_clean_high(d)
}
\arguments{
\item{d}{flagged BT tower data}
}
\description{
Removes NO,No2 and NOx values greater than 500 ppb and o3 greater than 300 ppb
Flags these as 5 - manually removed
}
