% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_BTT_EP_raw_processing.R
\name{BTT_EP_raw_data_processing}
\alias{BTT_EP_raw_data_processing}
\title{BTT_EP_raw_data_processing}
\usage{
BTT_EP_raw_data_processing(d, calibration_profile, extra_met, mgm3 = F)
}
\arguments{
\item{d}{Raw AQD Output}

\item{calibration_profile}{profile produced by BTT_calibration_profile from 1 min crit files}

\item{extra_met}{output of BTT_parse_1hz_met}

\item{mgm3}{if True, return NO and NO2 and mgm3 rather than a mixing ratio}
}
\value{
data.frame of the format to pass to Eddy Pro
}
\description{
Takes the raw output of the fast AQD NOx instruments at the BTT, and when supplied a calibration profile, output
by BTT_calibration_profile and the processed met data from the Reading data downloader - returns a file ready to be run
in Eddy Pro. Built ontop of the orginal script by Adam R. Vaughan for BTT(2012) and the Beijing Site(2017)
}
\author{
Will S. Drysdale
}
