% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_findcalrange_BTT.R
\name{BTT_find_cal_ranges}
\alias{BTT_find_cal_ranges}
\title{BTT Find Cal range}
\usage{
BTT_find_cal_ranges(d)
}
\arguments{
\item{d}{Bound Crit files}
}
\value{
data frame containg start and end times and rows for calibrations
}
\description{
Examines when the cal valve is on and off and returns the range of times that the cals occured in
used within BTT Calibration profile function
}
