% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_parse_1hz_met.R
\name{BTT_parse_1hz_met}
\alias{BTT_parse_1hz_met}
\title{BTT Parse 1 Hz Met}
\usage{
BTT_parse_1hz_met(d)
}
\arguments{
\item{d}{*raw* BTT met data from Readings data downloader. Files can be bound together to run for periods greater than 1 week}
}
\value{
tidied dataframe
}
\description{
Transforms the met data from the Reading Data dowloader at the BTT into a format useable in BTT_raw_data_processing
}
