% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_BTT_raw_data_processing.R
\name{BTT_raw_data_processing}
\alias{BTT_raw_data_processing}
\title{BTT_raw_data_processing}
\usage{
BTT_raw_data_processing(d, calibration_profile, extra_met)
}
\arguments{
\item{d}{Raw AQD Output}

\item{calibration_profile}{profile produced by BTT_calibration_profile from 1 min crit files}

\item{extra_met}{*raw* BTT met data from Readings data downloader. Files can be bound together to run for periods greater than 1 week}
}
\value{
data.frame of the format to pass to custom EC code
}
\description{
Takes the raw output of the fast AQD NOx instruments at the BTT, and when supplied a calibration profile, output
by BTT_calibration_profile and the *raw* met data from the Reading data downloader - returns a file ready to be run
in custom EC code. Built ontop of the orginal script by Adam R. Vaughan for BTT(2012) and the Beijing Site(2017)
}
\author{
Will S. Drysdale
}
