% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_c_.R
\name{c_}
\alias{c_}
\title{c_}
\usage{
c_(...)
}
\arguments{
\item{...}{objects to be concatenated.}
}
\description{
Updated \code{c()} for binding POSIXct. Preserves timezone (againist normal \code{c()} behaviour). If all tz are the same, use that tz, else use UTC
}
\author{
W. S. Drysdale
}
