% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_calculate_mass_balance.R
\name{calculate_mass_balance}
\alias{calculate_mass_balance}
\title{calculate_mass_balance}
\usage{
calculate_mass_balance(upwind_leg, downwind_matrix, v_matrix, xrows, yrows,
  pollutant, grid_scale, alt_scale)
}
\arguments{
\item{upwind_leg}{trimmed upwind leg from mass balance pre origin}

\item{downwind_matrix}{dowwind pollutant origin matrix read in as a dataframe from .csv}

\item{xrows}{number of x grid lines calculated in mass balances script}

\item{yrows}{number of y grid lines calculated in mass balances script}

\item{pollutant}{name of pollutant (single only) in upwind_leg}

\item{grid_scale}{scale of map grid axis (northing or easting) set in mass balances script}

\item{alt_scale}{scale of altitude axis set in mass balances script}

\item{v_matirx}{v_c origin matrix read in as a dataframe from .csv}
}
\value{
value of mass balance calculation in mg s-1
}
\description{
Calculates a mass_balance value at the end of the mass balances script
}
\author{
Will S. Drysdale

Beth Nelson
}
