% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_china_cal_flags.R
\name{china_cal_flags}
\alias{china_cal_flags}
\title{BTT Cal Flags}
\usage{
china_cal_flags(d)
}
\arguments{
\item{d}{Raw AQD NOx output}
}
\description{
When given a dataframe of AQD BTT NOx data locates calibrations,
using this information, recalculates the sensitivites and CE using a linear drift
}
