% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_china_calibration_profile.R
\name{china_calibration_profile}
\alias{china_calibration_profile}
\title{China Calibraion Profile}
\usage{

  china_calibration_profile(output_file_path = "~/../Google Drive/PhD/Freya's Magic Files/CHina/",
  crit_file_folder = "minute_data", percentage_drift = 50)
}
\arguments{
\item{output_file_path}{Location containing the folder where all of the crit files are stored}

\item{crit_file_folder}{name of folder located within output_file_path contining crit files}

\item{percentage_drift}{to be supplied to BTT_cal_clean to limit drift of bad cals}
}
\description{
From the 1 min AQD crit file, a profile of the corrected calibration paramaters is produced
}
\author{
Will S. Drysdale
}
