% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_China_raw_data_processing.R
\name{China_raw_data_processing}
\alias{China_raw_data_processing}
\title{China_raw_data_processing}
\usage{
China_raw_data_processing(d, my_sens1, my_sens2, calibration_profile)
}
\arguments{
\item{d}{Raw AQD Output}

\item{my_sens1}{channel 1 sensitvity}

\item{my_sens2}{channel 2 sensitvity}

\item{calibration_profile}{calibration_profile from china_calibration_profile()}
}
\value{
data.frame of the format to pass to custom EC code
}
\description{
Takes the raw output of the fast AQD NOx instruments from China, and is supplied with values for 
channel 1 + 2 sensitivity and a calibration profile for ce returns a file ready to be run in custom EC code. 
Built ontop of the orginal script by Adam R. Vaughan for BTT(2012) and the Beijing Site(2017)
}
\author{
Will S. Drysdale
}
