% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_chooseday.R
\name{chooseday}
\alias{chooseday}
\title{Choose Day}
\usage{
chooseday(d, myyear, mymonth, myday)
}
\arguments{
\item{d}{dataframe that day is to be selected from. date time information must be POSIXct in column labled "date"}

\item{myyear}{4 digit year as integer}

\item{mymonth}{month as integer, months < 10 must be given as single digit}

\item{myday}{day as integer, days < 10 must be given as single digit}
}
\value{
dataframe containing all data from corresponding day
}
\description{
Given a dataframe, allows either a specific day, range or selction of days to be selected
}
