% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_choosedays.R
\name{choosedays}
\alias{choosedays}
\title{Choose days}
\usage{
choosedays(d, myday)
}
\arguments{
\item{d}{dataframe that day is to be selected from. date time information must be POSIXct in column labled "date"}

\item{mymonth}{days as a vector of integers, months < 10 must be given as single digit}
}
\value{
dataframe containing all data from corresponding days
}
\description{
When chooseday() is asked for multiple days, they are selected and combined into a new data frame for further processing
}
