% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_choosefileprompt.R
\name{choosefileprompt}
\alias{choosefileprompt}
\title{Choose file prompt}
\usage{
choosefileprompt(flist)
}
\arguments{
\item{flist}{vector of fiel names}
}
\description{
when given a vector of file names, asks the user to select the number of which file they desire to load
}
