% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_chooseyears.R
\name{chooseyears}
\alias{chooseyears}
\title{Choose years}
\usage{
chooseyears(d, myyear)
}
\arguments{
\item{d}{dataframe that day is to be selected from. date time information must be POSIXct in column labled "date"}

\item{myyear}{4 digit year as vector of integers}
}
\value{
dataframe containing all data from corresponding years
}
\description{
When chooseday() is asked for multiple years, they are selected and combined into a new data frame for further processing
}
