% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_convert_met_data.R
\name{convert_met_data}
\alias{convert_met_data}
\title{Convert Met Data}
\usage{
convert_met_data(d)
}
\arguments{
\item{d}{Dataframe containing u = "u", v = "vv", temperature = "sonic_temp"}
}
\value{
Dataframe with ws,wd and temp columns appended
}
\description{
Converts wind vectors u and v into windspeed and direction.
Also converts Temperature / Speed of sound to  / Celcius
}
