% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_convert_met_data_wrk.R
\name{convert_met_data_wrk}
\alias{convert_met_data_wrk}
\title{Convert Met Data Worker}
\usage{
convert_met_data_wrk(d)
}
\arguments{
\item{d}{dataframe row by row via adply}
}
\description{
Passes each line of the data frame first to have the wind vectors, then temperature converted
}
