% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_convert_sos_c.R
\name{convert_temp_sos_c}
\alias{convert_temp_sos_c}
\title{Convert Temperature}
\usage{
convert_temp_sos_c(d)
}
\arguments{
\item{dataframe}{row by row via adply}
}
\description{
Converts temperature from speed of sound to Celcius
}
