% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_convert_sos_c2.R
\name{convert_temp_sos_c_2}
\alias{convert_temp_sos_c_2}
\title{Convert Temperature 2}
\usage{
convert_temp_sos_c_2(d)
}
\arguments{
\item{d}{column of data}
}
\description{
Converts temperature from speed of sound to Celcius
}
