% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_convert_wind_vectors.R
\name{convert_wind_vectors}
\alias{convert_wind_vectors}
\title{Convert Wind Vectors}
\usage{
convert_wind_vectors(d)
}
\arguments{
\item{d}{dataframe row by row via adply}
}
\description{
when provided a dataframe row by row via adply with wind vectors 
u and v in columns labled u and vv (DAQfactory format) creates 
two new columns ws and wd
Note, output is from a gill windmaster pro,where u aligns north positive and v aligns WEST positive
}
