% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_coords_from_kml.R
\name{coords_from_kml}
\alias{coords_from_kml}
\title{coords_from_kml}
\usage{
coords_from_kml(kml_path)
}
\arguments{
\item{kml_path}{location of kml file}
}
\description{
Reads coordinates into a dataframe from a hysplit kml file
}
