% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_create_area_polygon.R
\name{create_area_polygon}
\alias{create_area_polygon}
\title{Create Area Polygon}
\usage{
create_area_polygon(
  top_line,
  bot_line,
  reverse = F,
  crs_args = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0"
)
}
\arguments{
\item{top_line}{SpatialLines object describing the top edge of the polygon}

\item{bot_line}{SpatialLines object describing the bot edge of the polygon}

\item{reverse}{Does the bot_line need to be orderd in reverse to describe the polygon correctly, default FALSE}

\item{crs_args}{supply the CRS arguments to describe the polygon, default "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0"}
}
\description{
USing XY coordinate files with lat and lon as the first two columns creates a polygon
}
\author{
Will S. Drysdale
}
