% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_create_bin.R
\name{create_bin}
\alias{create_bin}
\title{Create Bin}
\usage{
create_bin(d, col, breaks)
}
\arguments{
\item{d}{dataframe}

\item{col}{column of data to assign breaks to}

\item{breaks}{sequence of breaks to used for bins}
}
\description{
Create a column of data bins based on breaks supplied by user. upper end of break is used as bin
}
\author{
W. S. Drysdale and F. A. Squires
}
