% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_crop_matrix_to_sqaure.R
\name{crop_matix_to_square}
\alias{crop_matix_to_square}
\title{Crop Matrix to Square}
\usage{
crop_matix_to_square(mymatrix, trim_size = 100, pad = 0)
}
\arguments{
\item{mymatrix}{matrix to be cropped}

\item{trim_size}{dimentions to crop matrix}

\item{pad}{if matrix is to be padded}
}
\description{
Crop a matrix to size. Discards a border around the trim_size, preserving the center of the matrix \cr
If \code{mymatrix} is smaller than the \code{trim_size}, \code{mymatrix} will be padded with \code{pad} \cr
\cr
Only supports matrices that both dimentions are to be cropped or padded. will not crop one dimention and pad the other
}
\author{
W S Drysdale
}
