% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_decrease_matrix_resolution.R
\name{decrease_matrix_resolution}
\alias{decrease_matrix_resolution}
\title{decrease_matrix_resolution}
\usage{
decrease_matrix_resolution(mymatrix, factor = 10)
}
\description{
decrease_matrix_resolution
}
