% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_diurnal_ggplot_from_timeVariation.R
\name{diurnal_ggplot_from_timeVariation}
\alias{diurnal_ggplot_from_timeVariation}
\title{Diurnal ggplot from timeVariation}
\usage{
diurnal_ggplot_from_timeVariation(d)
}
\arguments{
\item{d}{day.hour dataframe from openair::timeVariation}
}
\value{
data frame ready for ggplot
}
\description{
Uses the day.hour output of openair::timeVariation and returns a ggplot version of the plot
}
\author{
W. S. Drysdale
}
