% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_unique_event_full.R
\name{exp_day_comparison}
\alias{exp_day_comparison}
\title{Unique Day Comparions Full}
\usage{
exp_day_comparison(d, dtraj, date = "24/04/2016", wbefore = 4, wafter = 4,
  dbefore, dafter, pol = "nox", period = "15 min", use_polcol = T)
}
\arguments{
\item{d}{Dataframe containing time series}

\item{dtraj}{back trajectories covering time period to be studied}

\item{date}{DD/MM/YYYY of unique day}

\item{wbefore}{Number of weeks before unique date}

\item{wafter}{number of weeks after unique date}

\item{dbefore}{number of days before unique date (If not supplied, equal to wbefore)}

\item{dafter}{number of days before unique date (If not supplied, equal to wafter)}

\item{pol}{single pollutant to perfom experiment on}

\item{period}{over which the time series has been averaged (used to pad data series)}

\item{use_polcol}{T/F whether to use colours returned by pollution colour for consistency, currently supports no, no2, nox}
}
\description{
Performs full day comparison experiment using exp_unique_event and exp_unique_event_day
and create a PDF of results
}
