% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_unique_event.R
\name{exp_unique_event}
\alias{exp_unique_event}
\title{Unique Event Experiment}
\usage{
exp_unique_event(d, unique_day, weeks_before, weeks_after, pol, period)
}
\arguments{
\item{d}{dataframe containing the time series to apply the experiment to}

\item{unique_day}{the date of the unique event in the format "2016/11/06"}

\item{weeks_before}{numerical number of weeks preeceeding to include average days from}

\item{weeks_after}{numerical number of weeks following to include average days from}

\item{pol}{pollutant to perform experiment on - only handles one pollutant at a time}
}
\value{
dataframe ready for timeVariation
}
\description{
Produces a diurnal of a given day, and provides a comparison to the same type of day for given weeks in the past and/or future
e.g. Sunday 6th November 2016 is compared to an average of the preceeding 4 and following 4 sundays
}
