% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_unique_event_days.R
\name{exp_unique_event_days}
\alias{exp_unique_event_days}
\title{Unique Event Experiment Days}
\usage{
exp_unique_event_days(d, unique_day, day_before, day_after, pol, period)
}
\arguments{
\item{d}{dataframe containing the time series to apply the experiment to}

\item{unique_day}{the date of the unique event in the format "2016/11/06"}

\item{day_before}{numerical number of days preeceeding to include average days from}

\item{day_after}{numerical number of days following to include average days from}

\item{pol}{pollutant to perform experiment on - only handles one pollutant at a time}
}
\value{
}
\description{
Variation on Unique Event Experiment
Produces a diurnal of a given day, and provides a comparison to the same type of day for given days in the past and/or future
e.g. Sunday 6th November 2016 is compared to an average of the preceeding 4 and following 4 days
}
