% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_footprint_contribution.R
\name{f_cont}
\alias{f_cont}
\title{Footprint contribution}
\usage{
f_cont(foot_mat, percent_cont = 80)
}
\arguments{
\item{foot_mat}{maxtrix of footprint contributions}

\item{percent_cont}{percentage (expressed as whole number) of target contribution}
}
\description{
calculates the threshold above which all footprint contributions sum to greater than a target percentage contribution
}
\author{
W S Drysdale
}
