% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_FAAM_core_nox_merge.R
\name{FAAM_core_nox_merge}
\alias{FAAM_core_nox_merge}
\title{FAAM core nox merge}
\usage{
FAAM_core_nox_merge(ncdf_filepath, processed_nox)
}
\arguments{
\item{ncdf_filepath}{location of the netcdf}

\item{processed_nox}{nox csv, date can still be in excel time under "TheTime"}
}
\description{
Merge nox into the core faam ncdf
}
\author{
W S Drysdale
}
