% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_faam_leg_hysplit_intersect.R
\name{faam_leg_hysplit_intersect}
\alias{faam_leg_hysplit_intersect}
\title{faam_leg_hysplit_intersect}
\usage{
faam_leg_hysplit_intersect(flight_leg, traj)
}
\arguments{
\item{flight_leg}{flight_range_subset data.frame}

\item{traj}{coords_from_kml data.frame}
}
\value{
dataframe[lon,lat]
}
\description{
Calculates the interesct between a flight leg from flight_range_subset and a hysplit trajectory from coords_from_kml
}
\author{
Will S Drysdale
}
