% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_faam_lon_lattoeast_north.R
\name{faam_lon_lat.to.east_north}
\alias{faam_lon_lat.to.east_north}
\title{FAAM lon lat to easting northing}
\usage{
faam_lon_lat.to.east_north(df, missing_data)
}
\arguments{
\item{df}{faam merge csv with date column}

\item{missing_data}{missing dat flag for this flight}
}
\value{
original datafram with easting and northing columns
}
\description{
Takes the lat lon information from a faam merge file and adds easting and northing columsn
}
\author{
Will S. Drysdale
}
