% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_findcalrange.R
\name{find_cal_ranges}
\alias{find_cal_ranges}
\title{Frind Cal range}
\usage{
find_cal_ranges(d)
}
\arguments{
\item{Raw}{AQD NOx output}
}
\description{
Examines when the cal valve is on and off and returns the range of times that the cals occured in
}
