% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_find_range.R
\name{find_ranges}
\alias{find_ranges}
\title{Find Range}
\usage{
find_ranges(d, col)
}
\arguments{
\item{d}{dataframe with data and time stamp stored as UNIX_TS or UNIX.TS}

\item{col}{numerical value for column containg valve values}
}
\value{
dataframe containing "start","startrow","end","endrow"
}
\description{
Used to return ranges of rows where a valve has switched from 0 to 1 or vice versa. Improves upon find_cal_range and find_zero_range
}
