% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_findzerorange.R
\name{find_zero_ranges}
\alias{find_zero_ranges}
\title{Find Zero range}
\usage{
find_zero_ranges(d)
}
\arguments{
\item{d}{Raw AQD NOx output}
}
\value{
data frame containg start and end times and rows for zeros
}
\description{
Examines when the zero valve is on and off and returns the range of times that the zero occured in
}
