% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_flagcols.R
\name{flagcols}
\alias{flagcols}
\title{QC Flag a Column}
\usage{
flagcols(d, cols)
}
\arguments{
\item{d}{dataframe containing data to be flagged}

\item{cols}{integer giving column number providing location of data to be flagged}
}
\value{
dataframe with the one specified column flagged with formatting
}
\description{
Applies flags laid out in applyflags() to an individual column and formats dataframe accordingly
}
