% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_flagdata.R
\name{flagdata}
\alias{flagdata}
\title{Flag Data}
\usage{
flagdata(d, cols)
}
\arguments{
\item{d}{dataframe containing data to be flagged}

\item{cols}{vector argument giving column numbers providing location of data to be flagged}
}
\value{
dataframe containg flagged data
}
\description{
Applies Flags to data as specified in applyflags() ti multiple columns
}
