% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_flight_range_subset.R
\name{flight_range_subset}
\alias{flight_range_subset}
\title{Flight range subset}
\usage{
flight_range_subset(flight_data, flight_sum)
}
\arguments{
\item{flight_data}{faam merge file with date field created}

\item{flight_sum}{faam flight summary read through read.faaa_flight_sum}
}
\description{
creates list of range events in a flight
}
\author{
W S Drysdale
}
