% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_preproccess_AQD.R
\name{flux_pre_process}
\alias{flux_pre_process}
\title{Flux Pre Processing}
\usage{
flux_pre_process(d, site = "bt")
}
\arguments{
\item{d}{5hz output file}

\item{site}{Either "bt" or "china"}
}
\description{
Given Raw DAQfactory output from the AQD instruments at the BT Tower or China sites, reprocesses the concentration using calibration data
for 5Hz and 1 min crit data. 5Hz is reformatted for input into the ERF code
}
