% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_gen_theme.R
\name{gen_theme}
\alias{gen_theme}
\title{General Theme}
\usage{
gen_theme(fill_col = "grey97", line_col = "black")
}
\arguments{
\item{fill_col}{controls plot background colour,default grey97}

\item{line_col}{controls plot line colour, default black}
}
\description{
General theme for use with ggplot
}
