% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_get_DOY.R
\name{get_DOY}
\alias{get_DOY}
\title{Get DOY}
\usage{
get_DOY(d)
}
\arguments{
\item{d}{a POSIXct Date}
}
\value{
DOY.decimal_day
}
\description{
from a POSIXct Date, return DOY.decimal_day
}
\author{
Will S. Drysdale
}
