% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_naei_map_options.R
\name{get_naei_map_options}
\alias{get_naei_map_options}
\title{Get NAEI map options}
\usage{
get_naei_map_options(url = "https://naei.beis.gov.uk/data/map-uk-das")
}
\arguments{
\item{url}{url of emissions map selection page}
}
\description{
get the list of species that are avaliable for download - used by \code{get_naei_map()}
and useful for the user to match exact names
}
\author{
W. S. Drysdale
}
