% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_naei_timeseries.R
\name{get_naei_timeseries}
\alias{get_naei_timeseries}
\title{Get NAEI Timeseries}
\usage{
get_naei_timeseries(
  species = "Nitrogen oxides (NOx expressed as NO2)",
  from = 1970,
  to = 2019,
  file_path = NULL
)
}
\arguments{
\item{species}{vector of pollutant names as listed by \code{get_naei_timeseries_options()}}

\item{from}{start year}

\item{file_path}{path to save downloaded data to. If NULL data will be downloaded to a tempfile and returned}

\item{end}{year}
}
\description{
Downloads NAEI emissions timeseries from naei.beis.gov.uk/data/data-selector
}
\author{
W. S. Drysdale
}
