% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_naei_timeseries_options.R
\name{get_naei_timeseries_options}
\alias{get_naei_timeseries_options}
\title{Get NAEI Timeseries Options}
\usage{
get_naei_timeseries_options()
}
\description{
Returns the air pollutants available in the timerseries datasets at naei.beis.gov.uk/data/data-selector
}
\author{
W. S. Drysdale
}
