% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_Annulus.R
\name{ggplot_Annulus}
\alias{ggplot_Annulus}
\title{ggplot Annulus}
\usage{
ggplot_Annulus(polarAnnulusData, legendName = "z", tickScale = 0.05)
}
\arguments{
\item{polarAnnulusData}{data from \code{openair::polarAnnulus} output}

\item{legendName}{name for legend}

\item{tickScale}{Tick width is calculated from the maximum value in the u direction times this factor. Default 0.05}
}
\description{
Take the data output of \code{openair::polarAnnulus} and plot using ggplot
}
\author{
W. S. Drysdale
}
