% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_polarPlot.R
\name{ggplot_polarPlot}
\alias{ggplot_polarPlot}
\title{ggplot polarPlot}
\usage{
ggplot_polarPlot(polarPlotData, legendName = "z")
}
\arguments{
\item{polarPlotData}{data from \code{openair::polarPlot} output}

\item{legendName}{name for legend}
}
\description{
Take the data output of \code{openair::polarPlot} and plot using ggplot
}
\author{
W. S. Drysdale
}
