% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_ggwindrose.R
\name{ggwindRose}
\alias{ggwindRose}
\title{Wind Rose (ggplot2)}
\usage{
ggwindRose(d, ws = "ws", wd = "wd", breaks = 16, nbin = 5,
  col = NULL, group = NULL)
}
\arguments{
\item{d}{data.frame containing winddata}

\item{ws}{column name for windspeed as character, default ws}

\item{wd}{column name for winddirection as character, default wd}

\item{breaks}{number of wedges to break rose into default 16 (22.5 degree bins)}

\item{nbin}{number of bins to separate wind speed into, default 5}

\item{col}{colours for bins to pass to scale_fill_manual. must be of length nbin}

\item{group}{parameter to pass to facet wrap. Use non-standard evaluation (unquoted) to pass to \code{facet_wrap()}}
}
\description{
A ggplot2 windrose
}
\author{
W. S. Drysdale
}
