% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_goodcalprompt.R
\name{good_cal_prompt}
\alias{good_cal_prompt}
\title{Good cal Prompt}
\usage{
good_cal_prompt(d)
}
\arguments{
\item{Raw}{AQD NOx output}
}
\description{
Prompt used to assertain where cals are good or bad, reuires user input
}
