% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_between.R
\name{\%b\%}
\alias{\%b\%}
\title{Between operator}
\usage{
x \%b\% rng
}
\arguments{
\item{lhs}{value or vector of values to comapre with rhs}

\item{rhs}{pair of values in a vector}
}
\description{
Is a value between a pair of values, adapted from https://stackoverflow.com/questions/19441092/how-can-i-create-an-infix-between-operator \cr
Exclusive of bounds, irespective of order of rhs. Will use max and min values of rhs \cr
}
