% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_group_by_bin.R
\name{group_by_bin}
\alias{group_by_bin}
\title{Group By Bin}
\usage{
group_by_bin(d, col, group_col, .f = mean, ...)
}
\arguments{
\item{d}{dataframe containg col and group_col}

\item{col}{name of column to group}

\item{group_col}{name of column to group by i.e. the bins}

\item{.f}{function to summarise by default mean}

\item{...}{additional parameters for .f}
}
\description{
Takes a parameter and summarises by a bin column, then join this column onto the timeseries
can be used to bin secondary parameters for things like geom_boxplot, where the fill aesthetic must be pre binned
}
\author{
W. S. Drysdale and F. A. Squires
}
