% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_increase_matrix_resolution.R
\name{increase_matrix_resolution}
\alias{increase_matrix_resolution}
\title{increase_matrix_resolution}
\usage{
increase_matrix_resolution(mymatrix, size)
}
\arguments{
\item{mymatrix}{input matrix}

\item{size}{how many times to repeat values row and column wise, 4 will return 16 copies of the value}
}
\description{
repeats matrix elements to artificially increase resolution
}
