% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_interpolate_by_index.R
\name{interpolate_by_index}
\alias{interpolate_by_index}
\title{interpolate_by_index}
\usage{
interpolate_by_index(df, index_start, index_end, interp_col, index = "date",
  f = zoo::na.approx, extend = F)
}
\arguments{
\item{index_start}{value of index at begining of index. index must be unique. length must equal \code{length(interp_col)}}

\item{index_end}{end of index value}

\item{interp_col}{names of column, or vector of names to interpolate over}

\item{index}{name of index column, default "date"}

\item{f}{function to use for interpolation, default \code{zoo::na.approx}}

\item{d}{dataframe to interpolate over}

\item{extrap}{logical. when true, na.locf is applied to the whole interpolation column after the initial interpolation, expand the first and last values to the whole column}
}
\description{
interpolates between a coloumn between a defined start and end index
}
\author{
W. S. Drysdale
}
