% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_interpolate_by_index.R
\name{interpolate_by_index_wrk}
\alias{interpolate_by_index_wrk}
\title{interpolate_by_index}
\usage{
interpolate_by_index_wrk(
  d,
  index_start,
  index_end,
  interp_col,
  index = "date",
  f = zoo::na.approx
)
}
\arguments{
\item{d}{dataframe to interpolate over}

\item{index_start}{value of index at begining of index. index must be unique. length must equal \code{length(interp_col)}}

\item{index_end}{end of index value}

\item{interp_col}{names of column, or vector of names to interpolate over}

\item{index}{name of index column, default "date"}

\item{f}{function to use for interpolation, default \code{zoo::na.approx}}
}
\description{
interpolates between a coloumn between a defined start and end index
}
