% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_latlon_to_bbox.R
\name{latlon_to_bbox}
\alias{latlon_to_bbox}
\title{latlon_to_bbox}
\usage{
latlon_to_bbox(latlon, res_in_km = 100)
}
\arguments{
\item{latlon}{lat lon pair as a vector}

\item{res_in_km}{rough resolution in kilometers}
}
\value{
bounding box of  c(lowerleftlon,lowerleftlat,upperrightlon,upperrightlat), which can be passed directly to \code{ggmap::get_stamenmap}
}
\description{
Converts a lat lon pair of coordinates to a roughly square bounding box
}
\author{
W. S. Drysdale
}
