% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_lazy_read_nasa_ames.R
\name{lazy_read_nasa_ames}
\alias{lazy_read_nasa_ames}
\title{Lazy Read NASA Ames}
\usage{
lazy_read_nasa_ames(file, start_date, date_column = 1, tz = "UTC")
}
\arguments{
\item{file}{path of NASA Ames File}

\item{start_date}{date where DOY begins}

\item{tz}{timezone for date_from_decimal_day}

\item{column}{where date is stored, default 1}
}
\value{
list where item one is the header text and item 2 is the data as a dataframe
}
\description{
Function to read NASA Ames files into R in a basic way
}
\author{
Will S. Drysdale
}
