% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_leg_theme.R
\name{leg_theme}
\alias{leg_theme}
\title{Legend Theme}
\usage{
leg_theme(fill_col = "grey97", line_col = "black", pos = "right")
}
\arguments{
\item{fill_col}{controls plot background colour,default grey97}

\item{line_col}{controls plot line colour, default black}

\item{pos}{Legend Postion, default right}
}
\description{
Basic formatting for ggplot legends
}
