% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_length_of_lat_long.R
\name{length_of_long}
\alias{length_of_long}
\title{Length of Latitude/Longditue}
\usage{
length_of_long(deg_lat, r = 6378137)
}
\arguments{
\item{deg_lat}{degrees latitude for distance calculation}

\item{r}{radius of earth, default 6378137 (\code{length_of_long()} only)}
}
\value{
Size of 1 degree in m
}
\description{
Based on the latitude, calculate the size of degree lat or long in meters \cr
using equations from https://en.wikipedia.org/wiki/Latitude#Length_of_a_degree_of_latitude \cr
and https://en.wikipedia.org/wiki/Longitude#Length_of_a_degree_of_longitude
}
\author{
W. S. Drysdale
}
