% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_lets_split.R
\name{lets_split}
\alias{lets_split}
\title{Splitting Fucntion}
\usage{
lets_split(df, variable)
}
\arguments{
\item{df}{dataframe of time series passed by previous subsetting function}

\item{variable}{column by which df should be split}
}
\value{
list of dataframes
}
\description{
Used to subset data in certain scenarios where subset() and cutdata() fell short
}
