% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_loadBTT.R
\name{load_BTT}
\alias{load_BTT}
\title{Load BT Tower Data}
\usage{
load_BTT()
}
\description{
Loads in BT tower data based on parameters in the btconfig.R file.
Voncentration should be in the subdirectory "conc" and meterological data
in the sub directory "met"
Defaults to C:/Users/username/Documents/R/win-library/3.3/wsdmiscr/extdata/btconfig.R
}
