% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_naei_layers.R
\name{load_naei_layers}
\alias{load_naei_layers}
\title{Load NAEI Layers}
\usage{
load_naei_layers(
  dir = "C:/Users/Will/Google Drive/PhD/site_BT Tower/NAEI Data/NAEI NO2 2014/"
)
}
\arguments{
\item{dir}{directory containing".asc" files}
}
\description{
Read the unzipped NAEI sector layers into a raster stack
}
\author{
W. S. Drysdale
}
