% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_loop_group.R
\name{loop_group}
\alias{loop_group}
\title{Loop Group}
\usage{
loop_group(df, by = NULL, grp = NULL)
}
\arguments{
\item{df}{data.frame contained grouped/binned data}

\item{by}{value of the increment between groups, if NULL an attempt is made to guess this value}

\item{grp}{column containg grouping data. if NULL the first group of the grouped data.frame will be used.}
}
\description{
Take a data.frame with grouped data i.e grouped by hour for a diurnal, and duplicate \cr
the first group at the end of the data.frame, with its grouping variable == max(group)+1
}
\author{
W. S. Drysdale
}
