% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_mask_extract_date.R
\name{mask_extract_date}
\alias{mask_extract_date}
\title{Extract Date using Mask}
\usage{
mask_extract_date(string, mask, tz = "UTC")
}
\arguments{
\item{string}{string to extract date from}

\item{mask}{defining where date is in string}

\item{tz}{timezone code to pass to \code{ymd_hm()}}
}
\description{
By specifiic year = yy or yyyy, month = mm, day = dd, hour = HH, month = MM
}
\author{
W. S. Drysdale
}
