% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_mass_balance_pre_origin.R
\name{mass_balance_pre_origin}
\alias{mass_balance_pre_origin}
\title{mass_balance_pre_origin}
\usage{
mass_balance_pre_origin(flight_data, flight_summary, downwind_runs, upwind_runs,
  missing_flag = -99999, pollutant = "ch4_ppb", unit = "ppb",
  pollutant_mass = 16042)
}
\arguments{
\item{flight_data}{faam merge file with date column}

\item{flight_summary}{flight sumamry read in by read.faam_fligh_sum}

\item{downwind_runs}{vector of numbers corresponding to the list entries produced by flight_range_subset() for downwind}

\item{upwind_runs}{vector of numbers corresponding to the list entries produced by flight_range_subset() for upwind}

\item{missing_flag}{flag for missing data default -99999}

\item{pollutant}{string or vector of strings containg pollutants. see faam_mixing_ratio_to_mgm3}

\item{unit}{string or vector of strings containg units. see faam_mixing_ratio_to_mgm3}

\item{pollutant_mass}{string or vector of strings containg pollutant mass. see faam_mixing_ratio_to_mgm3}
}
\value{
list containg the upwind and downwind data.frames
}
\description{
Process a faam merge file ready for gridding in origin for mass balance. must have date column
}
\author{
Will S. Drysdale

Beth Nelson
}
