% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_massPlot.R
\name{massPlot}
\alias{massPlot}
\title{Mass Plot}
\usage{
massPlot(d, pol, fn, backTraj, start, end, s)
}
\arguments{
\item{d}{dataframe being processed}

\item{pol}{pollutants within dataset supplied as vector}

\item{fn}{the name and/or path where the file shall be saved}

\item{backTraj}{A backTrajectory file, pre merged with concentration data}

\item{start}{string in format "dd/mm/yyyy" to be passed to selectBydate()}

\item{end}{string in format "dd/mm/yyyy" to be passed to selectBydate()}

\item{s}{TRUE or FALSE. Is the data suitable for creating seasonal plots?}
}
\value{
A PDF containing all of the plots specified during the running of the function
}
\description{
Upon being given a dataset - this function produces a series of Openair plots using the avalible data.
These include timeVariation, wind and pollution rose and summary plots. 
Output into one PDF. Data should be supplied to comply with Openair formatting, with the time stamp in a column labled "date", in a format parseable by lubridates ymd_hms() function
}
