% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_massPlotmet.R
\name{massPlotmet}
\alias{massPlotmet}
\title{Plot Met}
\usage{
massPlotmet(d, pol, s)
}
\arguments{
\item{d}{dataframe being processed by massPlot}

\item{pol}{pollutants within dataset}
}
\description{
Creates the plots that require Met data, windrose, pollutionrose and polarplot
Where relevaent these plots are created for each species specified in massPlot(pol)
}
