% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_summaryPlot.R
\name{massplotsummary}
\alias{massplotsummary}
\title{Summary Interaction}
\usage{
massplotsummary(d)
}
\arguments{
\item{d}{dataframe being processed by massPlot}
}
\description{
Collects information about the data set with respect to passing values to summaryPlot()
}
