% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_massPlottimeVar.R
\name{massPlottimeVar}
\alias{massPlottimeVar}
\title{Time Variation Intereaction Level 1}
\usage{
massPlottimeVar(d, pol, s)
}
\arguments{
\item{d}{dataframe being processed by massPlot}

\item{pol}{pollutants within dataset}
}
\description{
Questions the user whether all the pollutants originally passed to massPlot() should be plotted on the same timeVariation() graph.
Regardless of whether a plot is created in this step,timeVarselection() is called to create plots of subsets of pollutants
}
