% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_massPlottraj.R
\name{massPlottraj}
\alias{massPlottraj}
\title{Plot Back Trajectories}
\usage{
massPlottraj(backTraj, pol, start, end, s)
}
\arguments{
\item{backTraj}{A backTrajectory file, pre merged with concentration data}

\item{pol}{pollutants within dataset}

\item{start}{string in format "dd/mm/yyyy" to be passed to selectBydate()}

\item{end}{string in format "dd/mm/yyyy" to be passed to selectBydate()}

\item{season}{If true, the time series is of a length suitable for seasonality comparison}
}
\description{
Incomplete and currently not run when calling massPlot()
}
