% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_dictionary_querys.R
\name{mdq}
\alias{mdq}
\title{Molecule Dictionary Query}
\usage{
mdq(mol = NULL, property = NULL)
}
\arguments{
\item{mol}{molecular formula of molecule to be queried}

\item{property}{property of the molecule to be returned not required for "molecules" or "structure}

\item{dictionary}{moleclar dictionary to use, defualt loaded by molecule_dictionary()}
}
\value{
Corresponding value from molecule_dictionary
}
\description{
Querys the molecule dictionary, mdq() is a shorthand wrapper fucntion
}
\author{
W S. Drysdale
}
