% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_dictionary_querys.R
\name{mol_dict_query}
\alias{mol_dict_query}
\title{Molecule Dictionary Query}
\usage{
mol_dict_query(mol = NA, property = NA, dictionary = molecule_dictionary())
}
\arguments{
\item{mol}{molecular formula of molecule to be queried}

\item{property}{property of the molecule to be returned not required for "molecules" or "structure}

\item{dictionary}{moleclar dictionary to use, defualt loaded by molecule_dictionary()}
}
\value{
Corresponding value from molecule_dictionary
}
\description{
Querys the molecule dictionary, mdq() is a shorthand wrapper fucntion
}
\author{
W S. Drysdale
}
