% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_molecule_dictionary.R
\name{molecule_dictionary}
\alias{molecule_dictionary}
\title{Molecule Dictionary}
\usage{
molecule_dictionary()
}
\arguments{
\item{none}{}
}
\value{
list containg the dictionary
}
\description{
List of molecules with information about them stored in a query-able fashion
mw for molecular weight
mw_unit returns the until of the molecular weight
parse_str string that can be passsed to parse(text = ) for correct formatting
name full name of the molecule
comment any commments about the entry
... additional queryable items listed in comment
}
\author{
W S. Drysdale
}
