% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_ncas_filename.R
\name{ncas_filename}
\alias{ncas_filename}
\title{NCAS Filename}
\usage{
ncas_filename(d, instrument, location, version)
}
\arguments{
\item{d}{dataframe of data being named}

\item{instrument}{string name of instument}

\item{location}{string name of location}

\item{version}{integer or double version number}
}
\value{
string filename
}
\description{
Formats the filename for btnox_adddata() into the NCAS data archive format:
<instrument_name>_<platform_name>_<Date-Time>_<data_product>_<version>
}
