% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_nest_tabset.R
\name{nest_tabset}
\alias{nest_tabset}
\title{Nest Tabset}
\usage{
nest_tabset(listTree, idTree, depth, parse_function,
  tabset_type = "tabset tabset-pills")
}
\arguments{
\item{listTree}{A list or list of lists which defines the structure of the tabset. \cr 
Names of list levels will be used to name tabs. Double underscore ,"__", will be replaced with a space}

\item{idTree}{A list or list of list, of character strings. Used for div ids. Must be unique throught the whole markdown document.}

\item{depth}{What depth should the tabset start at. I.e how many "#" would the top level begin with}

\item{parse_function}{A function to handle whatever lies at the end of the lists. e.g. use print for ggplots.}

\item{tabset_type}{either "tabset" or "tabset tabset-pills" Note the lack of curly braces and full stops.}
}
\description{
Create nested tabsets in RMarkdown. Supply a nested list of objects,\cr
unique ids for these and a function to display them. Returns the html to \cr
be contained in a r chunk with results = "asis"
}
\author{
W. S. Drysdale
}
