% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_pad_time_series.R
\name{pad_time_series}
\alias{pad_time_series}
\title{Pad Time Series}
\usage{
pad_time_series(d, period)
}
\arguments{
\item{d}{data frame with a "date" of class POSIXct POSIXt}

\item{period}{regular time series period in seconds}
}
\description{
Pad a time series by joining it to a regular timestamp
}
\author{
W S Drysdale
}
