% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_padtimeseries.R
\name{padtimeseries}
\alias{padtimeseries}
\title{Pad Time Series}
\usage{
padtimeseries(d, start, end, period)
}
\arguments{
\item{d}{dataframe of current time series - Date should be of POSIXct format in coloumn with header "date"}

\item{start}{String of start date for new timeseries in format "YYYY/MM/DD"}

\item{end}{String of end date for new timeseries in format "YYYY/MM/DD"}

\item{period}{time intervals for new time series - These should create time that match current time series}
}
\value{
Dataframe of new timeseries
}
\description{
Merges a data set into a new continuous time series
}
