% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubridate_supplements.R
\name{parse_excel_date}
\alias{parse_excel_date}
\title{Lubridate Supplements}
\usage{
parse_excel_date(x, tz = "UTC", type = "windows")
}
\description{
Lubridate supllements lifted from github.com/wacl-york/waclr
}
