% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_plot_flight_legs.R
\name{plot_flight_legs}
\alias{plot_flight_legs}
\title{Plot flight legs}
\usage{
plot_flight_legs(flight_data, flight_sum_path, plot_range = T,
  maptype = "toner", output_file_path = "plots/",
  colour_by = c("nox_conc", "o3_teco", "alt_gin"), flight_map,
  missing_flag = -9999, zoom = 8, logplot = c(T, F, F),
  range_control = T, cols = viridisLite::plasma(10))
}
\arguments{
\item{flight_data}{faam merge file}

\item{flight_sum_path}{file path of flight summary}

\item{plot_range}{T/F plot subsections of flight}

\item{maptype}{string for getmap, toner}

\item{output_file_path}{location to save plots}

\item{colour_by}{vector of cloums to colour by, defualt "nox_conc","o3_teco","alt_gin"}

\item{flight_map}{supply a gg_map file to override the automatic map generation}

\item{missing_flag}{data flag to convert to NA}

\item{zoom}{adjust map zoom level}

\item{logplot}{vector of equal length to colour_by in order should these be on log scales deafault (T,F,F)}

\item{range_control}{default false, enable for colour bars to change between flight segments}

\item{cols}{colour bar info default viridisLite::plasma(10)}
}
\description{
creates plots of the flight and its sections cloured by supplied parameters. date must be supllied as POSIXct, in column "date"
}
\author{
W S Drysdale
}
