% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_points_to_line.R
\name{points_to_line}
\alias{points_to_line}
\title{Points to Line}
\usage{
points_to_line(data, long, lat, id_field = NULL, sort_field = NULL)
}
\arguments{
\item{data}{dataframe containing XY data}

\item{long}{name of longditude column}

\item{lat}{name of latitude column}
}
\description{
Converts XY points to a spatial lines object. Copied from https://rpubs.com/walkerke/points_to_line
}
