% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_pollutioncolour.R
\name{pollution_colour}
\alias{pollution_colour}
\title{Pollution Colour}
\usage{
pollution_colour(pol, polcol)
}
\arguments{
\item{pol}{Vector of pollutant names}

\item{polcol}{list of colours and pollutants - where the colour is the item name and the pollutant is the list item}
}
\description{
When supplied a list of colours and values to supply to pollutant arugments in openair
Return a vector of pollutant names
}
