% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prep_aircraft_for_nasa_ames.R
\name{prep_aircraft_final_for_nasa_ames}
\alias{prep_aircraft_final_for_nasa_ames}
\title{Prepare Aircraft Final for NASA Ames}
\usage{
prep_aircraft_final_for_nasa_ames(d_final, missing_flag = 99999,
  col_names = NULL, date_col = 1, date_parse = waclr::parse_excel_date,
  origin = NULL)
}
\arguments{
\item{d_final}{_final output of \code{wsdmiscr::process.aircraft2()}}

\item{missing_flag}{what should NAs be replaced with}

\item{col_names}{what should the columns be renamed too}

\item{date_col}{which column contains the datetime info}

\item{date_parse}{function to transform the current date into a POSIXct}

\item{origin}{does date_parse require an origin? include it here}
}
\description{
Takes the _final output of \code{wsdmiscr::process.aircraft2()} and preps it for \code{wsdmiscr::create_nasa_ames_1D()}
}
