% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_process_aircraft2.R
\name{process.aircraft2}
\alias{process.aircraft2}
\title{Process Aircraft Data2}
\usage{
process.aircraft2(fn, dir_out = "", flight_id = "XXXX")
}
\arguments{
\item{fn}{file to read}

\item{dir_out}{output directory}

\item{flight_id}{ID of flight}
}
\value{
Saves two csvs into the output directory "processed" full output and "quicklook" only containing
time and concetrations. returns a list containg these two dataframes
}
\description{
Reads in an AQD NOx aircraft file and runs aircraft_cal_flags, prodcuing quicklook and full outputs
Used on Files after flight C019
}
\author{
Will S. Drysdale

Freya A. Squires
}
