% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_process_BTT_flags.R
\name{process_BTT_flags}
\alias{process_BTT_flags}
\title{Process BTT Flags}
\usage{
process_BTT_flags(d, cols)
}
\arguments{
\item{d}{BTT flagged data set}
}
\description{
Applies the rules that each flag specifies to the data set
0 data not used - Used when value is NA
1 good data - the default unless specific paramaters are met
2 data below limit of detection 0-1 ppb
3 data less than zero but within zero range -1-0 ppb
4 data less than -1 ppb, to be considered erroneous and should be converted to NA manually
}
