% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_process_zeros.R
\name{process_zeros}
\alias{process_zeros}
\title{process_zeros}
\usage{
process_zeros(
  df,
  zero_flag = c("zero_valve_1", "zero_valve_2"),
  calc_col = c("CH1_Hz", "CH2_Hz"),
  out_col = c("CH1_Hz_zero", "CH2_Hz_zero"),
  state_length = 15,
  trim_beg = 5,
  trim_end = 5,
  f = median,
  interp = T,
  ...
)
}
\arguments{
\item{df}{dataframe containing both the flag and calc data}

\item{zero_flag}{name(s) of columns to use as the statechange flag}

\item{calc_col}{name(s) of columns to use as the data to calculate f from}

\item{state_length}{length of the state, so they can be selected from the rle result}

\item{trim_beg}{trim a number of records from the begining of the state change}

\item{trim_end}{trim a number of records from the end of the state change}

\item{f}{function uses to calculate the statistic. Default median}

\item{...}{additional arguments to pass to f}
}
\description{
Calculates the result of function f, default median, over a series of state changes of equal length, determined by rle
}
