% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_read_1d_ncdf.R
\name{read.1D_ncdf}
\alias{read.1D_ncdf}
\title{Read 1D netCDF}
\usage{
read.1D_ncdf(path, var_names, dim_name = "time")
}
\arguments{
\item{path}{path too ncdf}

\item{var_names}{if only specific variables are required, supply a vector here}

\item{dim_name}{name of dimention to read in by}
}
\description{
reads in a netecesf file as a dataframe, using 1 dimention only
}
