% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_readeddypro.R
\name{read.eddypro}
\alias{read.eddypro}
\title{Read Eddy Pro}
\usage{
read.eddypro(filepath)
}
\arguments{
\item{filepath}{location of Eddy pro output (csv)}
}
\description{
Reads Eddy pro outputs into R
}
