% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_read_faam_flight_sum.R
\name{read.faam_flight_sum}
\alias{read.faam_flight_sum}
\title{read.faam_flight_sum}
\usage{
read.faam_flight_sum(flight_sum_path)
}
\arguments{
\item{flight_sum_path}{file location of flight summary}
}
\description{
Read a FAAM flight summary file into a dataframe
}
\author{
Will S. Drysdale
}
