% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_faam_ncdf_csv.R
\name{read.faam_ncdf}
\alias{read.faam_ncdf}
\title{Faam NetCDF to csv}
\usage{
read.faam_ncdf(ncdf_filepath)
}
\arguments{
\item{ncdf_filepath}{location of the netCDF}
}
\value{
returns a .csv with the timestamp in the "date" column in POSIXct format. Also contains "UNIX_TS" unix timestamp and "SECS" seconds since midnight
}
\description{
Takes the NetCDF for FAAM flight data and returns the key inforamtion as a csv
}
\author{
W S Drysdale
}
