% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_read_1d_ncdf.R
\name{read_1D_ncdf}
\alias{read_1D_ncdf}
\title{Read 1D netCDF}
\usage{
read_1D_ncdf(path, var_names, dim_name = "time", tz = "UTC",
  origin = "1970-01-01")
}
\arguments{
\item{path}{path too ncdf}

\item{var_names}{if only specific variables are required, supply a vector here}

\item{dim_name}{name of dimention to read in by if dim_name is time, format is assumed to be seconds since 1970-01-01 \cr
override with \code{origin}}

\item{tz}{defaults to UTC}

\item{origin}{defaults to 1970-01-01}
}
\description{
reads in a netecesf file as a dataframe, using 1 dimention only
}
