% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_read_fdt.R
\name{read_fdt}
\alias{read_fdt}
\title{Read Float Data Format}
\usage{
read_fdt(filename, headers = NULL)
}
\arguments{
\item{filename}{path to .fdt file}

\item{headers}{character vector of header names, or data.frame where names(headers) returns a character vector of header names}
}
\value{
data.frame
}
\description{
Read the .fdt output of PTRwid doi: 10.5194/amt-8-3903-2015
}
