% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_read_nasa_ames.R
\name{read_nasa_ames}
\alias{read_nasa_ames}
\title{Read NASA Ames}
\usage{
read_nasa_ames(file, SComm_header = FALSE, header_length = NULL, ...)
}
\arguments{
\item{file}{path to the file}

\item{SComm_header}{Logical default false. If the final line of special comments contains the data header, set to true}

\item{header_length}{by default the header length is obtained from the first line of the file. Supply a value here to override this.}
}
\description{
Reads a NASA Ames file by removing the header and parsing the rest of the data using read.table()
}
\author{
W. S. Drysdale
}
