% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_reshape_pan.R
\name{reshape_pan}
\alias{reshape_pan}
\title{Reshape PAN}
\usage{
reshape_pan(d, ecd)
}
\arguments{
\item{d}{data frame of PAN data}

\item{ecd}{either 1 or 2, depending on ECD channel}
}
\value{
long PAN data
}
\description{
Takes a dataframe of PAN data, which is in a pseudo wide format and \cr
reshapes it into long, based of know lags relative to the time stamp \cr
Input data should be in the form: \cr
1. Date (DMY) \cr
2. Time (HMS) \cr
... \cr
7. PAN1A (or PAN1B) \cr
8. PAN2A (or PAN2B) \cr
9. PAN3A (or PAN3B) \cr
10. PAN1A_ppb (or PAN1B_ppb) \cr
11. PAN2A_ppb (or PAN2B_ppb) \cr
12. PAN3A_ppb (or PAN3B_ppb) \cr
Columns 3 - 6 inclusive are not used but expected
}
\author{
W S. Drysdale
}
