% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_savefileprompt.R
\name{savefileprompt}
\alias{savefileprompt}
\title{Save File Prompt}
\usage{
savefileprompt(d, fn)
}
\arguments{
\item{d}{dataframe being saved}

\item{fn}{filename determined by ncas_filename()}
}
\description{
Manages the top level flow when saving the file from btnox_adddata()
}
