% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_selectpol.R
\name{selectpol}
\alias{selectpol}
\title{Select Pollutant}
\usage{
selectpol(n)
}
\arguments{
\item{n}{vector of column headers in dataframe - for error checking}
}
\value{
vector of pollutant names
}
\description{
Collects pollutant information to plot for timeVar
}
