% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_selectpol1.R
\name{spstep1}
\alias{spstep1}
\title{Select Pollutant Sub step 1}
\usage{
spstep1(n)
}
\arguments{
\item{n}{vector of column headers}
}
\description{
Allows for error catching whilst selecting pollutants
}
