% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_selectpol2.R
\name{spstep2}
\alias{spstep2}
\title{Select Pollutant Sub step 2}
\usage{
spstep2(n, t, x, i)
}
\arguments{
\item{n}{vector of column headers}

\item{t}{number of pollutants to plot}

\item{x}{first pollutant selected in step 1}

\item{i}{count variable from loop in selectpol}
}
\description{
Allows for error catching whilst selecting multiple pollutants
}
