% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_mean_line.R
\name{stat_mean_line}
\alias{stat_mean_line}
\title{Stat Mean Line}
\usage{
stat_mean_line(mapping = NULL, data = NULL, geom = "hline",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{x}{x}

\item{y}{y}
}
\description{
Plots a mean of all data in the facet in ggplot, taken from here https://stackoverflow.com/questions/46327431/
}
