% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_mean_line.R
\docType{data}
\name{StatMeanLine}
\alias{StatMeanLine}
\title{stat_mean_line}
\format{An object of class \code{StatMeanLine} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.}
\usage{
StatMeanLine
}
\arguments{
\item{y}{}
}
\description{
Plots a mean of all data in the facet in ggplot, taken from here https://stackoverflow.com/questions/46327431/
}
\keyword{datasets}
