% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sub_all.R
\name{sub_all}
\alias{sub_all}
\title{Subset All}
\usage{
sub_all(d)
}
\arguments{
\item{d}{Dataframe of time series, date time information must be in column labled "date". Date information will be processed using lubridate's ymd_hms, make sure the format is compatible}
}
\value{
Large list object with sublists ultimatley containing data frames with mixtures of the subsetting parameters applied
}
\description{
Takes a time series and divides it into lists of various paramaeters: Year, Month, Season, Weekday/end,Night/day
}
\examples{
sub_all(mydata)

}
