% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sub_lowlevel.R
\name{sub_lowlevel}
\alias{sub_lowlevel}
\title{Low Level Subsetting}
\usage{
sub_lowlevel(d)
}
\arguments{
\item{d}{dataframe of time series passed by previous subsetting function}
}
\value{
List contaning dataframes and lists of all the low level subsetting combinations
}
\description{
Applies all subsetting below month level to a given dataframe: Weekday/end and Night/day
}
