% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sub_month.R
\name{sub_month}
\alias{sub_month}
\title{Subset Month}
\usage{
sub_month(d)
}
\arguments{
\item{d}{dataframe of time series passed by previous subsetting function}
}
\value{
Large list containing data frames of subsetted months and lists with lower level subsetting applied
}
\description{
Subsets the data into months, and then applies lower level subsets additionally
}
