% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sub_nightday.R
\name{sub_nightday}
\alias{sub_nightday}
\title{Subset Night/Day}
\usage{
sub_nightday(d)
}
\arguments{
\item{d}{dataframe of time series passed by previous subsetting function}
}
\value{
list of two dataframes containg subsetting by day or night time
}
\description{
Subsets data based on whether the data was recorded during the day or night
}
