% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sub_season.R
\name{sub_season}
\alias{sub_season}
\title{subset Season}
\usage{
sub_season(d, season)
}
\arguments{
\item{d}{dataframe of time series passed by previous subsetting function}

\item{season}{character string "spring","summer,"winter","autumn" determining which season is returned}
}
\value{
dataframe of requested season
}
\description{
Subsets into seasons spring: MAM, summer: JJA, autumn: SON, winter: DJF
}
