% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sub_seasonlevel.R
\name{sub_seasonlevel}
\alias{sub_seasonlevel}
\title{Season Level Subsetting}
\usage{
sub_seasonlevel(d)
}
\arguments{
\item{d}{dataframe of time series passed by previous subsetting function}
}
\value{
Large list containing data frames of subsetted seasons and lists with lower level subsetting applied
}
\description{
Applies sub_season() for each season, and then applies lower level subsets additionally
}
