% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sub_weekdaygrp.R
\name{sub_weekdaygrp}
\alias{sub_weekdaygrp}
\title{Subset Weekdays together}
\usage{
sub_weekdaygrp(d)
}
\arguments{
\item{d}{dataframe of time series passed by previous subsetting function}
}
\value{
dataframe with weekdays only
}
\description{
Forms subsets of weekdays combined
}
