% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sub_weekdays.R
\name{sub_weekdays}
\alias{sub_weekdays}
\title{Subset Weekdays individually}
\usage{
sub_weekdays(d, dn)
}
\arguments{
\item{d}{dataframe of time series passed by previous subsetting function}

\item{dn}{boolean value to determine whether the function should apply the Night/day subsetting}
}
\value{
either Data Frames for weekdays separately or lists for each day with Night/Day subsetting applied
}
\description{
Forms subsets of weekdays separatley, with lower level subsetting avaliable
}
