% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sub_weekendgrp.R
\name{sub_weekendgrp}
\alias{sub_weekendgrp}
\title{Subset Weekends together}
\usage{
sub_weekendgrp(d)
}
\arguments{
\item{d}{dataframe of time series passed by previous subsetting function}
}
\value{
dataframe containing data from Saturday and Sundays
}
\description{
Forms subset containing data from both Saturdays and Sundays
}
