% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sub_weekends.R
\name{sub_weekends}
\alias{sub_weekends}
\title{Subset weekends individually}
\usage{
sub_weekends(d, dn)
}
\arguments{
\item{d}{dataframe of time series passed by previous subsetting function}

\item{dn}{boolean value to determine whether the function should apply the Night/day subsetting}
}
\value{
either Data Frames for Saturday and Sunday separately or lists for each day with Night/Day subsetting applied
}
\description{
Forms subsets of Saturdays and Sundays separatley, with lower level subsetting avaliable
}
