% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sub_yearlevel.R
\name{sub_yearlevel}
\alias{sub_yearlevel}
\title{Year Level Subsetting}
\usage{
sub_yearlevel(d)
}
\arguments{
\item{d}{dataframe of time series passed by previous subsetting function}
}
\value{
Large list containing dataframes of years and lists with lower level subsetting applied
}
\description{
Subsets the data into years, and then applies lower level subsets additionally
}
