% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_subset_grid.R
\name{subset_grid}
\alias{subset_grid}
\title{Subset Grid}
\usage{
subset_grid(grid, sp_poly)
}
\arguments{
\item{grid}{grid file e.g NAEI as SpatialGridDataFrame}

\item{sp_poly}{polygon defining area to subset}
}
\description{
Select an area of a grid file using a SpatialPolygons object
}
\author{
Will S. Drysdale
}
