% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_summarise.acf.r
\name{summarise.acf}
\alias{summarise.acf}
\title{Summarise ACF}
\usage{
summarise.acf(acf, .f = median, freq = 5, ...)
}
\arguments{
\item{acf}{acf output from e4r}

\item{.f}{function to summarise by, default mean}

\item{freq}{data frequency}

\item{...}{additonal arguments to pass to .f}
}
\description{
Summarise the ACF output by a function (e.g median) per scalar
}
\author{
W. S. Drysdale
}
