% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_time_series_compare.R
\name{time_series_compare}
\alias{time_series_compare}
\title{Time Series Compare}
\usage{
time_series_compare(df, date = "date", s1_col = "F_NO1_mass",
  s2_col = "no_flux_mg_m2_h", s1_name = "eddy4R", s2_name = "EddyPro",
  comparison_name = bquote("NO Flux / mg " ~ m^-2 ~ h^-1), time_res = 3600,
  font = "Garamond", scale = 5, den_range = c(-5, 30))
}
\arguments{
\item{df}{dataframe containing the two time series to compare}

\item{date}{column name for timestamp}

\item{s1_col}{series 1 column name}

\item{s2_col}{series 2 column name}

\item{s1_name}{series 1 name (method, instrument etc),used on plots}

\item{s2_name}{series 2 name (method, instrument etc),used on plots}

\item{comparison_name}{what about the time series is being comapred, used on plots}

\item{time_res}{time resolution of data in seconds}

\item{font}{select font for plotting}

\item{scale}{scales the axis tick and axis label text}

\item{den_range}{for density plotting, overwhat range of densities should the overlap integral be calculated c(min,max)}
}
\value{
List object containing list of plots and dataframe of stats
}
\description{
Produces basic plots and stats summary comparing two time series
}
\author{
W S Drysdale
}
