% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_timeVarselection.R
\name{timeVarselection}
\alias{timeVarselection}
\title{Time Variation Intereaction Level 2}
\usage{
timeVarselection(d, pol, s)
}
\arguments{
\item{d}{dataframe being processed by massPlot}

\item{pol}{pollutants within dataset}
}
\description{
Allows the user to specify a list of pollutants to be plotted on the same timeVariation plot
To Skip this or to break out of the plotting cycle, the user should answer "n" to the intitial question
}
