% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_vec_col_attr.R
\name{vec_col_attr}
\alias{vec_col_attr}
\title{Vector of Column Attributes}
\usage{
vec_col_attr(x, which)
}
\arguments{
\item{x}{dataframe}

\item{which}{attribute to be returned}
}
\description{
returns a character vector of a specified attribute for each column in a dataframe
}
\author{
W S. Drysdale
}
