% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_workingdirectoryprompt.R
\name{workingdirectoryprompt}
\alias{workingdirectoryprompt}
\title{Working Directory Prompt}
\usage{
workingdirectoryprompt(d, fn)
}
\arguments{
\item{d}{dataframe being saved}

\item{fn}{filename determined by ncas_filename()}
}
\description{
Manages the bottom level flow when saving the file from btnox_adddata()
}
