% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_meta.R
\name{zip_map}
\alias{zip_map}
\title{zip_map}
\usage{
zip_map(path, .f = utils::read.csv, map_f = purrr::map, zipMeta = NULL)
}
\arguments{
\item{path}{path to zip file, to pass to}

\item{.f}{function to map, default \code{utils::read.csv}}

\item{map_f}{describes how \code{.f} will be mapped.}

\item{zipMeta}{automatically obtained via \code{zipMeta = unzip(path,list = T)} unless an alternative is specified. 
Files to be read must be stored in \code{zipMeta$Name}}
}
\description{
map a function across the contents of a zip file. Default reads each element using \code{read.csv} into a list created by \code{purrr::map}
}
\author{
W. S. Drysdale
}
