% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_applyflags.R
\name{applyflags}
\alias{applyflags}
\title{Apply Flags}
\usage{
applyflags(d, cols, flagnum, LOD)
}
\arguments{
\item{d}{dataframe containing data to be flagged}

\item{cols}{integer value of column containg data to be flagged}

\item{flagnum}{integer value of column where flags are to be stored}
}
\value{
dataframe with QC flag column added
}
\description{
Adds a column with QC flags based on the paramaters listed:
Flag Concentration data:
0 good data - the default unless specific paramaters are met
1 data below limit of detection 0-LOD
3 NA
}
