% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_findcalrange.R
\name{find_cal_ranges}
\alias{find_cal_ranges}
\title{Find Cal range}
\usage{
find_cal_ranges(d)
}
\arguments{
\item{d}{Raw AQD NOx output}
}
\value{
data frame containg start and end times and rows for calibrations
}
\description{
Examines when the cal valve is on and off and returns the range of times that the cals occured in
}
