% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_extract_inventory_value.R
\name{footprint_inventory_extract}
\alias{footprint_inventory_extract}
\title{Footprint Inventory Extract}
\usage{
footprint_inventory_extract(fp, inv, loc, full_output = F)
}
\arguments{
\item{fp}{footprint matrix}

\item{inv}{inventory as a projected raster object}

\item{loc}{c(lat,long) coordinates for the center of the footprint matrix}

\item{full_output}{default false, when true a spatial points dataframe of the inventory values and footprint are returned}
}
\value{
The sum of the inventory weighted by the footprint
}
\description{
Extracts the values of an inventory, weighted by their flux footprint
}
\author{
W S Drysdale
}
