% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_naei_map.R
\name{get_NAEI_map}
\alias{get_NAEI_map}
\title{Get NAEI map}
\usage{
get_NAEI_map(species = "Nitrogen oxides (NOx expressed as NO2)")
}
\arguments{
\item{species}{name of species of interest as written the in dropdown box at \url{naei.beis.gov.uk/data/map-uk-das}}
}
\description{
Download the NAEI UK emissions map
}
\author{
W. S. Drysdale
}
