% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_plot_flight_legs.R
\name{plot_flight_legs}
\alias{plot_flight_legs}
\title{Plot flight legs}
\usage{
plot_flight_legs(flight_data, flight_sum_path, plot_range = T,
  maptype = "toner", output_file_path, colour_by = c("nox_conc", "o3_teco",
  "alt_gin"), flight_map, missing_flag, ...)
}
\arguments{
\item{flight_data}{faam merge file}

\item{flight_sum_path}{file path of flight summary}

\item{plot_range}{T/F plot subsections of flight}

\item{maptype}{string for getmap, toner}

\item{output_file_path}{location to save plots}

\item{flight_map}{supply a gg_map file to override the automatic map generation}

\item{missing_flag}{data flag to convert to NA}
}
\description{
creates plots of the flight and its sections cloured by supplied parameters. date must be supllied as POSIXct, in column "date"
}
\author{
W S Drysdale
}
