% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ndist.R
\name{plot_ndist}
\alias{plot_ndist}
\title{Plot Normal distribution}
\usage{
plot_ndist(df, columns = c(CH1_Hz, CH2_Hz))
}
\arguments{
\item{df}{data.frame containing a date column as POSIXct and the columns to plot}

\item{columns}{column names expressed unquoted.}
}
\description{
Fit and plot a normal distribution histogram alongside a timeseries of the data the distribution has been fitted from
}
\author{
W. S. Drysdale
}
