% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_read_e4r.R
\name{read.e4r}
\alias{read.e4r}
\title{Read eddy4R}
\usage{
read.e4r(file_path_to_EC, grp, dfdd = "2017-01-01", dur = 3600,
  round = "hour", tz = "UTC")
}
\arguments{
\item{file_path_to_EC}{folder containing eddy4r output}

\item{grp}{name this group of files e.g run_id}

\item{dfdd}{date as "yyyy-mm-dd" for date_from_decimal_day}

\item{dur}{flux aggregation period in seconds}

\item{round}{sensible value to round the flux aggregation period to}
}
\value{
A data.frame containing the bound mean, error, stationarity and standard deviation files. With the timestamp adjusted to the begining of the period
}
\description{
Reads the eddy4r output of Will's branch as of 18/11/20
}
