% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_mean_line.R
\docType{data}
\name{stat_mean_line}
\alias{stat_mean_line}
\alias{StatMeanLine}
\title{Stat Mean Line}
\format{
An object of class \code{StatMeanLine} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatMeanLine
}
\description{
Plots a mean of all data in the facet in ggplot, taken from here https://stackoverflow.com/questions/46327431/
}
\keyword{datasets}
